transfer off logic from SearchActivity.kt
